refactor: move block-workflows endpoint to keeperhub/api/
Move implementation to keeperhub/api/internal/block-workflows/ and
leave a thin re-export in app/api/internal/block-workflows/ to follow
the custom code policy for upstream fork separation.
